+++ tests/training/test_nudged_source.sh
@@ -7,6 +7,6 @@ gsutil -m rm -r $OUTPUT
 
 python -m fv3net.regression.sklearn \
     $TRAINING_DATA \
-    tests/sklearn_regression/train_sklearn_model_nudged_source.yaml  \
+    tests/training/train_sklearn_model_nudged_source.yaml  \
     $OUTPUT \
     --no-train-subdir-append