+++ workflows/prognostic_run_diags/load_diagnostic_data.py
@@ -25,6 +25,7 @@ DIM_RENAME_INVERSE_MAP = {
 }
 VARNAME_SUFFIX_TO_REMOVE = ["_coarse"]
 _DIAG_OUTPUT_LOADERS = []
+MASK_VARNAME = "SLMSKsfc"
 
 
 def _adjust_tile_range(ds: xr.Dataset) -> xr.Dataset:
@@ -189,7 +190,7 @@ def load_verification(
 def _load_standardized(path):
     logger.info(f"Loading and standardizing {path}")
     m = fsspec.get_mapper(path)
-    ds = xr.open_zarr(m, consolidated=True).load()
+    ds = xr.open_zarr(m, consolidated=True)
     return standardize_gfsphysics_diagnostics(ds)
 
 
@@ -234,7 +235,10 @@ def load_dycore(url: str, grid_spec: str, catalog: intake.Catalog) -> DiagArg:
     # open verification
     logger.info("Opening verification data")
     verification_c48 = load_verification(
-        ["40day_c384_atmos_8xdaily"], catalog, coarsening_factor=8, area=grid_c384.area
+        ["40day_c384_atmos_8xdaily_may2020"],
+        catalog,
+        coarsening_factor=8,
+        area=grid_c384.area,
     )
 
     # open prognostic run data
@@ -262,17 +266,25 @@ def load_physics(url: str, grid_spec: str, catalog: intake.Catalog) -> DiagArg:
 
     # open grid
     logger.info("Opening Grid Spec")
-    # grid_c384 = standardize_gfsphysics_diagnostics(vcm.open_tiles(grid_spec))
+    grid_c384 = standardize_gfsphysics_diagnostics(vcm.open_tiles(grid_spec))
     grid_c48 = standardize_gfsphysics_diagnostics(catalog["grid/c48"].to_dask())
 
     # open verification
-    # TODO: load physics diagnostics for SHiELD data. Currently slow due to chunking.
-    verification_c48 = xr.Dataset()
+    verification_c48 = load_verification(
+        ["40day_c384_diags_time_avg_may2020"],
+        catalog,
+        coarsening_factor=8,
+        area=grid_c384.area,
+    )
     verification_c48 = add_derived.physics_variables(verification_c48)
 
     # open prognostic run data
     logger.info(f"Opening prognostic run data at {url}")
     prognostic_output = _load_prognostic_run_physics_output(url)
     prognostic_output = add_derived.physics_variables(prognostic_output)
+
+    # Add mask information if not present
+    if MASK_VARNAME in prognostic_output and MASK_VARNAME not in verification_c48:
+        verification_c48[MASK_VARNAME] = prognostic_output[MASK_VARNAME].copy()
 
     return prognostic_output, verification_c48, grid_c48