+++ workflows/prognostic_run_diags/test_generate_report.py
@@ -1,8 +1,9 @@
-from generate_report import upload, _parse_metadata
+from generate_report import upload, _parse_metadata, detect_rundirs
 
 import pytest
-from google.cloud.storage.client import Client
+import fsspec
 import uuid
+from google.cloud.storage.client import Client
 
 
 @pytest.fixture()
@@ -38,3 +39,30 @@ def test__parse_metadata():
     run = "blah-blah-baseline"
     out = _parse_metadata(run)
     assert out == {"run": run, "baseline": "Baseline", "one_step": "blah-blah"}
+
+
+def test_detect_rundirs(tmpdir):
+
+    fs = fsspec.filesystem("file")
+
+    rundirs = ["rundir1", "rundir2"]
+    for rdir in rundirs:
+        tmpdir.mkdir(rdir).join("diags.nc").write("foobar")
+
+    tmpdir.mkdir("not_a_rundir").join("useless_file.txt").write("useless!")
+
+    result = detect_rundirs(tmpdir, fs)
+
+    assert len(result) == 2
+    for found_dir in result:
+        assert found_dir in rundirs
+
+
+def test_detect_rundirs_fail_less_than_2(tmpdir):
+
+    fs = fsspec.filesystem("file")
+
+    tmpdir.mkdir("rundir1").join("diags.nc").write("foobar")
+
+    with pytest.raises(ValueError):
+        detect_rundirs(tmpdir, fs)